refactor: ThemeSection 컴포넌트 이동
