+++ src/pages/MainPage.tsx
@@ -1,7 +1,7 @@
 import Page from '@components/templates/Page';
 import Container from '@components/atoms/container/Container';
 import MainBanner from '@components/organisms/main/MainBanner';
-import ThemeSection from '@components/organisms/main/ThemeSection';
+import ThemeSection from '@components/organisms/main/theme/ThemeSection';
 import AIButtonSection from '@components/organisms/main/AIButtonSection';
 import RankingSection from '@components/organisms/main/ranking/RankingSection';
 import { useState } from 'react';
